requirements.txt
alembic==1.0.5
bz2file==0.98
Click==7.0
cycler==0.10.0
Django==2.1.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.1.0
nibabel==2.3.3
numpy==1.16.0
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
WTForms==2.2.1
